requirements.txt
Flask==2.3.3
PyQt5==5.15.9
qrcode[pil]==7.4.2
Pillow==10.0.1